Added significant functionality for retry/backoff with jitter calculations and multi-statuscode retry strategies.
